+++ clinica/pipelines/deeplearning_prepare_data/deeplearning_prepare_data_cli.py
@@ -49,10 +49,11 @@ class DeepLearningPrepareDataCLI(ce.CmdParser):
             "extract_method",
             help="""Format of the extracted features. Three options:
             'image' to convert to PyTorch tensor the complete 3D image,
-            'patch' to extract 3D volumetric patches and
-            'slice' to extract 2D slices from the image.
+            'patch' to extract 3D volumetric patches,
+            'slice' to extract 2D slices from the image and
+            'roi' to extract region of interest.
             By default the features are extracted from the cropped image.""",
-            choices=["image", "slice", "patch"],
+            choices=["image", "slice", "patch", "roi"],
             default="image",
         )
 
@@ -107,6 +108,40 @@ class DeepLearningPrepareDataCLI(ce.CmdParser):
                 single channel (default: --slice_mode rgb).""",
             choices=["rgb", "single"],
             default="rgb",
+        )
+
+        optional_roi = self._args.add_argument_group(
+            f"{Fore.BLUE}Pipeline options if you chose 'roi' extraction{Fore.RESET}"
+        )
+        optional_roi.add_argument(
+            "--roi_list",
+            help="List of regions to be extracted",
+            type=str,
+            nargs="+",
+            default=None,
+        )
+        optional_roi.add_argument(
+            "--roi_uncrop_output",
+            help="Disable cropping option so the output tensors have the same size than the whole image.",
+            action="store_true",
+            default=False,
+        )
+
+        optional_roi.add_argument(
+            "-ct",
+            "--custom_template",
+            help="""Name of the template used when modality is set to custom.""",
+            type=str,
+            default=None,
+        )
+        optional_roi.add_argument(
+            "-cmp",
+            "--custom_mask_pattern",
+            help="""If given will select only the masks containing the string given.
+                    The mask with the shortest name is taken.
+                    This argument is taken into account only when the modality is custom.""",
+            type=str,
+            default="",
         )
 
         optional_pet = self._args.add_argument_group(
@@ -168,6 +203,10 @@ class DeepLearningPrepareDataCLI(ce.CmdParser):
             "stride_size": args.stride_size,
             "slice_direction": args.slice_direction,
             "slice_mode": args.slice_mode,
+            "roi_list": args.roi_list,
+            "roi_uncrop_output": args.roi_uncrop_output,
+            "custom_template": args.custom_template,
+            "custom_mask_pattern": args.custom_mask_pattern,
             "use_uncropped_image": args.use_uncropped_image,
             "custom_suffix": args.custom_suffix,
             "acq_label": args.acq_label,